+++ client/package.json
@@ -17,7 +17,6 @@
     "@types/react-dom": "^18.0.1",
     "antd": "^4.19.5",
     "axios": "^0.26.1",
-    "babel-plugin-macros": "^3.1.0",
     "i18next": "^21.6.16",
     "i18next-browser-languagedetector": "^6.1.4",
     "prop-types": "^15.8.1",
