+++ client/src/Pages/Profile.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import UserInfoCard from '../components/UserInfoCard/UserInfoCard';
+import UserInfoCard from '../Components/UserInfoCard';
 
 function Profile() {
   const userInfo = {
